Rename features and identities
The old naming mixed apples and oranges with names like l2-ipv4-acl.
IPv4 is an instance of a L3 technology, so keeping to oranges the
correct name is eth-ipv4-acl, which is precisely what I've renamed
things to.

I've removed "mixed" as I find it superfluous. I've also added in a new
type called eth-ipv46.

Descriptions are updated, I think it's an improvement in wording but it
could possible taking further improvements :)

I renamed tcp/udp/icmp-acl to -fields instead as they are not separate
ACL types but rather support for matching certain header fields.

Should icmp-fields be split into ICMPv4 and ICMPv6? They are different
after all..