Use choice statements for ACE validity
We introduce a choice statement per OSI (or TCP/IP) layer so that we
enforce a single choice of matches per ACE. For example, it does not
make any sense to match on an IPv4 source address and a IPv6 destination
address in the same ACE, thus we prevent such a configuration by putting
ipv4-acl and ipv6-acl in a choice statement.

Use git diff -w (to ignore whitespace) to see that it's really a minimal
change.